+++ app/charts/table/cell-desktop.tsx
@@ -1,4 +1,4 @@
-import { Box, Theme } from "@mui/material";
+import { Box, Theme, Typography } from "@mui/material";
 import { makeStyles } from "@mui/styles";
 import { hcl } from "d3-color";
 import { ScaleLinear } from "d3-scale";
@@ -10,6 +10,7 @@ import { LinkedCellWrapper } from "@/charts/table/linked-cell-wrapper";
 import { ColumnMeta, TableChartState } from "@/charts/table/table-state";
 import { Tag } from "@/charts/table/tag";
 import { Flex } from "@/components/flex";
+import { OverflowTooltip } from "@/components/overflow-tooltip";
 import { Observation } from "@/domain/data";
 
 const useStyles = makeStyles((theme: Theme) => ({
@@ -61,12 +62,15 @@ export const CellDesktop = ({
     barShowBackground,
   } = columnMeta;
   const classes = useStyles();
-  const { links } = useChartState() as TableChartState;
+  const { links, shouldApplyWidthLimits } = useChartState() as TableChartState;
 
   switch (columnMeta.type) {
     case "text":
+      const textContent = columnMeta.formatter(cell);
+
       return (
         <Flex
+          {...cell.getCellProps()}
           sx={{
             alignItems: "center",
             justifyContent:
@@ -78,10 +82,24 @@ export const CellDesktop = ({
             fontWeight: textStyle,
             px: 3,
           }}
-          {...cell.getCellProps()}
         >
           <LinkedCellWrapper cell={cell} columnMeta={columnMeta} links={links}>
-            {columnMeta.formatter(cell)}
+            {shouldApplyWidthLimits ? (
+              <OverflowTooltip arrow title={textContent}>
+                <Typography
+                  component="span"
+                  variant="inherit"
+                  noWrap
+                  sx={{ lineHeight: 1.5 }}
+                >
+                  {textContent}
+                </Typography>
+              </OverflowTooltip>
+            ) : (
+              <Box component="span" sx={{ lineHeight: 1.5 }}>
+                {textContent}
+              </Box>
+            )}
           </LinkedCellWrapper>
         </Flex>
       );
@@ -89,8 +107,8 @@ export const CellDesktop = ({
       const { colorScale: cColorScale } = columnMeta;
       return (
         <Flex
-          sx={{ alignItems: "center", fontWeight: textStyle, pl: 1, pr: 3 }}
           {...cell.getCellProps()}
+          sx={{ alignItems: "center", fontWeight: textStyle, pl: 1, pr: 3 }}
         >
           <LinkedCellWrapper cell={cell} columnMeta={columnMeta} links={links}>
             <Tag tagColor={cColorScale(cell.value)}>
@@ -128,6 +146,7 @@ export const CellDesktop = ({
           sx={{
             flexDirection: "column",
             justifyContent: "center",
+            alignItems: "flex-end",
             // Padding is a constant accounted for in the
             // widthScale domain (see table state).
             px: `${BAR_CELL_PADDING}px`,