+++ app/charts/table/width-limit.ts
@@ -1,0 +1,5 @@
+import { ColumnMeta } from "@/charts/table/table-state";
+
+export const columnCanBeWidthLimited = (columnType: ColumnMeta["type"]) => {
+  return columnType === "text" || columnType === "category";
+};