+++ app/utils/chart-config/constants.ts
@@ -1,3 +1,3 @@
-export const CONFIGURATOR_STATE_VERSION = "5.1.0";
+export const CONFIGURATOR_STATE_VERSION = "5.2.0";
 
-export const CHART_CONFIG_VERSION = "5.1.0";
+export const CHART_CONFIG_VERSION = "5.2.0";